chore(SE-62): relocate interfaces from services to constants folder
