+++ app/dashboard/(user)/_components/ProfileCard.tsx
@@ -3,7 +3,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import Link from "next/link";
-import { UserProfile } from "@/services/profileService";
+import { UserProfile } from "@/constants/Users";
 
 interface ProfileCardProps {
   className?: string;
