+++ constants/Users.ts
@@ -14,3 +14,36 @@ export type TenantType = {
   businessName: string;
   registerDate: string;
 };
+
+export interface UserProfile {
+  id: string;
+  email: string;
+  firstName: string;
+  lastName: string;
+  phoneNumber?: string;
+  avatarUrl?: string;
+  joinedAt: Date;
+  userType: string;
+  tenantInfo?: TenantProfile;
+}
+
+export interface TenantProfile {
+  businessName: string;
+  taxId?: string;
+  registeredDate: Date;
+}
+
+export interface UserImage {
+  avatarUrl: string;
+}
+
+export interface UserImageResponse {
+  data: UserImage;
+}
+
+export interface UpdateProfile {
+  firstName: string;
+  lastName: string;
+  phoneNumber?: string;
+  avatar?: string;
+}
